token expiration

See merge request videoflix-49808/videoflix_backend!24

+++ videoflix_app/views.py
@@ -26,7 +26,8 @@ from rest_framework.permissions import IsAuthenticated
 from .models import UserContinueWatchVideo, UserFavoriteVideo, Video
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.authentication import TokenAuthentication
-from .serializers import VideoSerializer
+from datetime import timedelta
+from django.utils import timezone
 User = get_user_model()
 
 # CACHETTL = getattr(settings, 'CACHETTL', DEFAULT_TIMEOUT)
@@ -78,6 +79,7 @@ def activate_user(request, uidb64, token):
 
 class RequestPasswordReset(APIView):
     permission_classes = [AllowAny]
+    TokenAuthentication = [AllowAny]
     User = get_user_model()
     serializer_class = ResetPasswordRequestSerializer
 
@@ -123,21 +125,37 @@ class RequestPasswordReset(APIView):
         else:
             return Response({"error": "User not found"}, status=status.HTTP_404_NOT_FOUND)
         
-class PasswordResetView(generics.GenericAPIView):
+class PasswordResetView(APIView):
     permission_classes = []
-    def post(self, request, token):
-        
+
+    def get(self, request, token):
         reset_obj = PasswordReset.objects.filter(token=token).first()
-        print(f'reset_obj {reset_obj}')
+        if not reset_obj:
+            return Response({'error': 'Invalid token'}, status=status.HTTP_400_BAD_REQUEST)
+
+        token_lifetime = timedelta(hours=24)
+        print(token_lifetime)
+        if timezone.now() > reset_obj.created_at + token_lifetime:
+            return Response({'error': 'Token expired'}, status=status.HTTP_400_BAD_REQUEST)
+
+        return Response({'success': 'Token is valid'}, status=status.HTTP_200_OK)
+
+    def post(self, request, token):
+        reset_obj = PasswordReset.objects.filter(token=token).first()
         if not reset_obj:
             return Response({'error': 'Invalid token'}, status=400)
+
+        token_lifetime = timedelta(hours=24)
+        if timezone.now() > reset_obj.created_at + token_lifetime:
+            return Response({'error': 'Token expired'}, status=400)
+
         user = User.objects.filter(email=reset_obj.email).first()
         if user:
             user.set_password(request.data['password'])
             user.save()
             reset_obj.delete()
             return Response({'success': 'Password updated'})
-        else: 
+        else:
             return Response({'error': 'No user found'}, status=404)
 
 @api_view(['GET', 'POST'])
